BP editor type 1 (v2)
Improved the edit-ability of the input fields in the editor. Now, only
the qualitative schemas are not accessible for editing once the tables
for the specification of the qualitative value sets have been created.
This is the min possible integrity requirement with max possible
freedom for editing of other elements.